requirements.txt
pytest-asyncio==1.3.0
pytest-playwright==0.7.1
requests==2.32.5